♻️  [#93] Refactor: product-review-modal 폴더 구조 변경
