+++ src/widgets/product-review-modal/components/image-upload/ReviewImageUpload.tsx
@@ -1,18 +1,20 @@
 import { UseFormSetValue } from 'react-hook-form'
 import { ImageInput } from '@shared/ui'
 import { createImageUpload } from '@shared/utils'
-import type { CreateReview } from '@shared/types'
+import type { CreateReview, UpdateReview } from '@shared/types'
 
-interface ProductReviewImageUploadProps extends Pick<CreateReview, 'images'> {
-  setValue: UseFormSetValue<CreateReview>
+interface ReviewImageUploadProps extends Pick<CreateReview, 'images'> {
+  setValue: UseFormSetValue<any>
   imageIndex: number
+  previewImage?: string
 }
 
-export default function ProductReviewImageUpload({
+export default function ReviewImageUpload({
   setValue,
   images,
   imageIndex,
-}: ProductReviewImageUploadProps) {
+  previewImage,
+}: ReviewImageUploadProps) {
   const onUploadSuccess = async (file: File, index: number) => {
     try {
       const { url } = await createImageUpload(file)
@@ -22,6 +24,7 @@ export default function ProductReviewImageUpload({
   }
 
   const onCancel = (index: number) => {
+    console.log(index)
     const filteredImage = images.filter((image) => image.index !== index)
     setValue('images', filteredImage)
   }
@@ -32,6 +35,7 @@ export default function ProductReviewImageUpload({
       onCancel={onCancel}
       onSuccess={onUploadSuccess}
       className="tablet:w-[135px] tablet:h-[135px] mobile:w-[140px] mobile:h-[140px] flex-shrink-0"
+      previewImage={previewImage || ''}
     />
   )
 }
